# gateway/event.swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
      "/conditions": {
        "post": {
          "operationId": "ConditionsController_create",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateConditionDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "조건이 성공적으로 생성됨",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConditionResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "잘못된 요청",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            }
          },
          "summary": "조건 생성",
          "tags": [
            "conditions"
          ]
        },
        "get": {
          "operationId": "ConditionsController_findAll",
          "parameters": [],
          "responses": {
            "200": {
              "description": "모든 조건 목록",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConditionsResponseDto"
                  }
                }
              }
            },
            "500": {
              "description": "서버 오류",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            }
          },
          "summary": "모든 조건 조회",
          "tags": [
            "conditions"
          ]
        }
      },
      "/conditions/{id}": {
        "get": {
          "operationId": "ConditionsController_findOne",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "description": "조건 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "조건 정보",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConditionResponseDto"
                  }
                }
              }
            },
            "404": {
              "description": "조건을 찾을 수 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            }
          },
          "summary": "특정 조건 조회",
          "tags": [
            "conditions"
          ]
        },
        "put": {
          "operationId": "ConditionsController_update",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "description": "조건 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateConditionDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "업데이트된 조건 정보",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConditionResponseDto"
                  }
                }
              }
            },
            "400": {
              "description": "잘못된 요청",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            },
            "404": {
              "description": "조건을 찾을 수 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            }
          },
          "summary": "조건 업데이트",
          "tags": [
            "conditions"
          ]
        },
        "delete": {
          "operationId": "ConditionsController_remove",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "description": "조건 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "삭제된 조건 정보",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ConditionResponseDto"
                  }
                }
              }
            },
            "404": {
              "description": "조건을 찾을 수 없음",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponseSwaggerDto"
                  }
                }
              }
            }
          },
          "summary": "조건 삭제",
          "tags": [
            "conditions"
          ]
        }
      },
...
      "/reward-history/user/{userId}": {
        "get": {
          "operationId": "RewardHistoryController_findByUserId",
          "parameters": [
            {
              "name": "userId",
              "required": true,
              "in": "path",
              "description": "사용자 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "사용자별 보상 이력 목록",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/RewardHistoryResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "사용자별 보상 이력 조회",
          "tags": [
            "reward-history"
          ]
        }
      },
...
      "/user-event-rewards/user/{userId}": {
        "get": {
          "operationId": "UserEventRewardsController_findAllByUserId",
          "parameters": [
            {
              "name": "userId",
              "required": true,
              "in": "path",
              "description": "유저 ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "유저의 모든 이벤트 보상 요청 목록을 반환합니다.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/UserEventRewardRequestResponseDto"
                    }
                  }
                }
              }
            }
          },
          "summary": "유저의 모든 이벤트 보상 요청 조회",
          "tags": [
            "user-event-rewards"
          ]
        }
      },
...
    },
    "info": {
      "title": "이벤트 서비스 API",
      "description": "이벤트 관리 시스템을 위한 API 문서",
      "version": "1.0",
      "contact": {
  
      }
    },
...
    "servers": [],
    "components": {
      "securitySchemes": {
        "JWT": {
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "type": "http",
          "name": "JWT",
          "description": "JWT 토큰을 입력하세요",
          "in": "header"
        }
      },
      "schemas": {
...
        "ConditionResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "조건 ID",
              "example": "60a12d5c8f45e32b4c9cde1f"
            },
            "category": {
              "type": "string",
              "description": "조건 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "subType": {
              "type": "string",
              "description": "조건 하위 타입",
              "enum": [
                "total_days",
                "after_signup",
                "invite_success",
                "meso_spent",
                "cash_topup",
                "item_collect",
                "quest_clear",
                "return_after_days"
              ],
              "example": "total_days"
            },
            "target": {
              "type": "object",
              "description": "조건 기준값 (예: 3일, 아이템 ID 등)",
              "example": 3
            },
            "description": {
              "type": "string",
              "description": "조건 설명",
              "example": "3일 연속 출석 시 보상 지급"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 시간",
              "example": "2025-05-20T00:00:00.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 시간",
              "example": "2025-05-20T00:00:00.000Z"
            }
          },
          "required": [
            "id",
            "category",
            "subType",
            "createdAt",
            "updatedAt"
          ]
        },
        "RewardResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "보상 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "type": {
              "type": "string",
              "description": "보상 타입",
              "enum": [
                "item",
                "meso",
                "mPoint"
              ],
              "example": "item"
            },
            "quantity": {
              "type": "number",
              "description": "보상 수량",
              "example": 1
            },
            "itemId": {
              "type": "string",
              "description": "아이템 ID (타입이 아이템인 경우에만 존재)",
              "example": "60d21b4667d0d8992e610c85"
            },
            "description": {
              "type": "string",
              "description": "보상 설명",
              "example": "레어 아이템 보상"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 일시",
              "example": "2025-05-20T00:37:58.000Z"
            }
          },
          "required": [
            "id",
            "type",
            "quantity",
            "createdAt",
            "updatedAt"
          ]
        },
...
        "CreateConditionDto": {
          "type": "object",
          "properties": {
            "category": {
              "type": "string",
              "description": "조건 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "subType": {
              "type": "string",
              "description": "조건 하위 타입",
              "enum": [
                "total_days",
                "after_signup",
                "invite_success",
                "meso_spent",
                "cash_topup",
                "item_collect",
                "quest_clear",
                "return_after_days"
              ],
              "example": "total_days"
            },
            "target": {
              "type": "string",
              "description": "조건 기준값 (예: 3일, 아이템 ID 등)",
              "example": "3"
            },
            "description": {
              "type": "string",
              "description": "조건 설명",
              "example": "3일 연속 출석 시 보상 지급"
            }
          },
          "required": [
            "category",
            "subType"
          ]
        },
        "ErrorResponseSwaggerDto": {
          "type": "object",
          "properties": {
            "code": {
              "type": "string",
              "description": "에러 코드",
              "example": "CONDITION_NOT_FOUND"
            },
            "message": {
              "type": "string",
              "description": "에러 메시지",
              "example": "조건을 찾을 수 없습니다"
            },
            "timestamp": {
              "type": "string",
              "description": "에러 발생 시간",
              "example": "2025-05-20T00:09:55.000Z"
            }
          },
          "required": [
            "code",
            "message",
            "timestamp"
          ]
        },
        "ConditionsResponseDto": {
          "type": "object",
          "properties": {
            "conditions": {
              "description": "조건 목록",
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ConditionResponseDto"
              }
            },
            "total": {
              "type": "number",
              "description": "총 조건 수",
              "example": 10
            }
          },
          "required": [
            "conditions",
            "total"
          ]
        },
        "UpdateConditionDto": {
          "type": "object",
          "properties": {
            "category": {
              "type": "string",
              "description": "조건 카테고리",
              "enum": [
                "attendance",
                "social",
                "economy",
                "mission",
                "special"
              ],
              "example": "attendance"
            },
            "subType": {
              "type": "string",
              "description": "조건 하위 타입",
              "enum": [
                "total_days",
                "after_signup",
                "invite_success",
                "meso_spent",
                "cash_topup",
                "item_collect",
                "quest_clear",
                "return_after_days"
              ],
              "example": "total_days"
            },
            "target": {
              "type": "string",
              "description": "조건 기준값 (예: 3일, 아이템 ID 등)",
              "example": "3"
            },
            "description": {
              "type": "string",
              "description": "조건 설명",
              "example": "3일 연속 출석 시 보상 지급"
            }
          }
        },
...
        "RewardHistoryResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "보상 이력 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "userId": {
              "type": "string",
              "description": "사용자 ID",
              "example": "user123"
            },
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "60d21b4667d0d8992e610c85"
            },
            "reward": {
              "description": "보상 목록",
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/RewardResponseDto"
              }
            },
            "claimedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수령 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "deliveryType": {
              "type": "string",
              "description": "지급 방식",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "example": "manual_claim"
            },
            "status": {
              "type": "string",
              "description": "보상 상태",
              "enum": [
                "pending",
                "success",
                "failed"
              ],
              "example": "success"
            },
            "reason": {
              "type": "string",
              "description": "실패 이유 (실패 시에만 존재)",
              "example": "이미 수령한 보상입니다"
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 일시",
              "example": "2025-05-20T00:37:58.000Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 일시",
              "example": "2025-05-20T00:37:58.000Z"
            }
          },
          "required": [
            "id",
            "userId",
            "eventId",
            "reward",
            "claimedAt",
            "deliveryType",
            "status",
            "createdAt",
            "updatedAt"
          ]
        },
...
        "UserEventRewardRequestResponseDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "요청 ID",
              "example": "60d21b4667d0d8992e610c90"
            },
            "userId": {
              "type": "string",
              "description": "유저 ID",
              "example": "user-123"
            },
            "eventId": {
              "type": "string",
              "description": "이벤트 ID",
              "example": "event-001"
            },
            "rewardId": {
              "type": "string",
              "description": "보상 ID",
              "example": "60d21b4667d0d8992e610c87"
            },
            "trigger": {
              "type": "string",
              "description": "트리거 타입",
              "enum": [
                "manual",
                "auto"
              ],
              "example": "manual"
            },
            "deliveryType": {
              "type": "string",
              "description": "보상 전달 타입",
              "enum": [
                "immediate",
                "manual_claim",
                "scheduled"
              ],
              "example": "manual_claim"
            },
            "rewardSnapshot": {
              "type": "object",
              "description": "보상 스냅샷 정보",
              "example": {
                "type": "item",
                "quantity": 1,
                "itemId": "item-001",
                "description": "레어 아이템"
              }
            },
            "status": {
              "type": "string",
              "description": "보상 상태",
              "enum": [
                "pending",
                "success",
                "failed"
              ],
              "example": "pending"
            },
            "requestedAt": {
              "format": "date-time",
              "type": "string",
              "description": "유저가 보상을 요청한 시간",
              "example": "2025-01-01T12:00:00Z"
            },
            "reason": {
              "type": "string",
              "description": "보상 처리 결과에 대한 이유 (실패 시 오류 메시지 등)",
              "example": "이미 보상을 받았습니다."
            },
            "createdAt": {
              "format": "date-time",
              "type": "string",
              "description": "생성 시간",
              "example": "2025-01-01T12:00:00Z"
            },
            "updatedAt": {
              "format": "date-time",
              "type": "string",
              "description": "수정 시간",
              "example": "2025-01-01T12:00:00Z"
            }
          },
          "required": [
            "id",
            "userId",
            "eventId",
            "rewardId",
            "trigger",
            "deliveryType",
            "rewardSnapshot",
            "status",
            "requestedAt",
            "reason",
            "createdAt",
            "updatedAt"
          ]
        }
      }
    }
  }